add RHEL7 to OS enumerations, convert internal logging to use net-snmp logging, change lcoation of tmp files from / to /var/crash.

+++ recorder/crash.c
@@ -45,15 +45,15 @@
 
 #define SUCCESSFULLY_PROCESSED_TAGFILE ".hp-ams"
 #define DEBUG_TRIGGER_FILE "/tmp/.hp-ams"
-#define DEBUG_LOG_FILE "/tmp/hp-ams.log"
 
 #define LINVER_UNK         0x0
 #define LINVER_RHEL_UNK    0x1
 #define LINVER_RHEL_5      0x2
 #define LINVER_RHEL_6      0x3
-#define LINVER_SLES_UNK    0x4
-#define LINVER_SLES_10     0x5
-#define LINVER_SLES_11     0x6f
+#define LINVER_RHEL_7      0x4
+#define LINVER_SLES_UNK    0x5
+#define LINVER_SLES_10     0x6
+#define LINVER_SLES_11     0x7
 
 typedef struct os_dep_config_t {
     int linux_version;
@@ -76,7 +76,6 @@ os_dep_config dep;
 /* global values */
 extern int s_ams_rec_class;
 extern int s_ams_rec_handle;
-FILE* debug_output = NULL;
 int debug = 0;
 
 /* some external prototypes we need */
@@ -117,37 +116,6 @@ static void strip(char *buf, int sz) {
     }
 }
 
-static void debug_log(const char *fmt, ...) {
-    va_list argp;
-    pid_t pid = getpid();
-
-
-    /*
-    time_t t = time(NULL);
-    char t_str[ = ctime(&t);
-    */
-
-    if (debug) {
-        if (NULL == debug_output) {
-            debug_output = fopen(DEBUG_LOG_FILE, "a");
-            assert(NULL != debug_output);
-        }
-        //fprintf(debug_output, "[%s]: ", ctime(&t));
-        fprintf(debug_output, "hpHelper[%d] ", pid);
-        va_start(argp, fmt);
-        vfprintf(debug_output, fmt, argp); 
-        fprintf(debug_output, "\n");
-        va_end(argp);
-        fflush(debug_output);
-    }
-}
-
-static void close_debug_log(void) {
-    if (NULL != debug_output) {
-        fclose(debug_output);
-    }
-}
-
 static void get_os_deps(os_dep_config* dep) {
     int linux_version = LINVER_UNK;
     distro_id_t* distro = NULL;
@@ -157,7 +125,9 @@ static void get_os_deps(os_dep_config* dep) {
     distro = getDistroInfo();
     if (NULL != distro) {
         if (memcmp(rh_string, distro->Vendor, strlen(rh_string)) == 0) {
-            if (6 == distro->Version) {
+            if (7 == distro->Version) {
+                linux_version = LINVER_RHEL_7;
+            } else if (6 == distro->Version) {
                 linux_version = LINVER_RHEL_6;
             } else if (5 == distro->Version) {
                 linux_version = LINVER_RHEL_5;
@@ -181,6 +151,7 @@ static void get_os_deps(os_dep_config* dep) {
     case LINVER_RHEL_UNK:
     case LINVER_RHEL_5:
     case LINVER_RHEL_6:
+    case LINVER_RHEL_7:
     case LINVER_SLES_UNK:
     case LINVER_SLES_10:
     case LINVER_SLES_11:
@@ -257,7 +228,8 @@ static int gettmpname(char *buf, int bufsz) {
     int rc = 0;
 
     memset(buf, 0, bufsz);
-    strncpy(buf, dep.output_template, bufsz-1);
+    strncpy(buf, dep.crash_dir, bufsz-1);
+    strncat(buf, dep.output_template, bufsz-1);
     
     fd = mkstemp(buf);
     
@@ -265,7 +237,8 @@ static int gettmpname(char *buf, int bufsz) {
         rc = fd;
     } else {
         close(fd);
-        unlink(buf);
+        if (debug = 0)
+            unlink(buf);
     }
     return rc;
 }
@@ -629,8 +602,8 @@ static int get_path_value_from_file(char* filename, char* buf, int bufsz) {
                 if (memcmp(line, match, match_sz-1) == 0) {
                     char *nptr = line+match_sz;
 
-                    debug_log("Found path configuration %s in %s on line %d.",
-                              nptr, filename, lcount);
+                    DEBUGMSGTL(("rec:crash","Found path configuration %s in %s on line %d.",
+                              nptr, filename, lcount));
                     strncpy(buf, nptr, bufsz-1);
                     rc = 0;
                 }
@@ -670,19 +643,19 @@ static int get_crash_dir(char *buf, int bufsz) {
                 /* failed to obtain a path from the kdump config, use
                  * default
                  */
-                debug_log("Did not find path value in %s, using default value: %s", dep.kdump_conf, dep.crash_dir);
+                DEBUGMSGTL(("rec:crash","Did not find path value in %s, using default value: %s", dep.kdump_conf, dep.crash_dir));
                 strncpy(buf, dep.crash_dir, bufsz-1);
             } else {
                 /* obtained a path from the kdump config, copy it into
                  * the result buffer
                  */
-                debug_log("Found path value in %s, using: %s", dep.kdump_conf, buffer);
+                DEBUGMSGTL(("rec:crash","Found path value in %s, using: %s", dep.kdump_conf, buffer));
                 strncpy(buf, buffer, bufsz-1);
             }
 
         } else {
             /* no kdump.conf, use default */
-            debug_log("No %s file found, using default: %s", dep.kdump_conf, dep.crash_dir);
+            DEBUGMSGTL(("rec:crash","No %s file found, using default: %s", dep.kdump_conf, dep.crash_dir));
             strncpy(buf, dep.crash_dir, bufsz-1);
         }
     } else {
@@ -751,7 +724,7 @@ static llist* find_crash_entries(int previous_flag) {
     get_crash_dir(crash_dir, PATH_MAX);
 
     if (does_exist(crash_dir)) {
-        debug_log("Crash directory exists (%s), examining for candidates to process.", crash_dir);
+        DEBUGMSGTL(("rec:crash","Crash directory exists (%s), examining for candidates to process.", crash_dir));
         dp = opendir(crash_dir);
         if (NULL != dp) {
             while ((ep = readdir(dp))) {
@@ -773,18 +746,18 @@ static llist* find_crash_entries(int previous_flag) {
                     if (previous_flag && !stat_result) {
                         // looking for previous, and found one
                         l = insert_item(ep->d_name, l);
-                        debug_log("Adding previously processed: %s/%s.", crash_dir, ep->d_name);
+                        DEBUGMSGTL(("rec:crash", "Adding previously processed: %s/%s.", crash_dir, ep->d_name));
                     } else {
                         if (!previous_flag && stat_result) {
                             // looking for new, and did not find touch file
                             l = insert_item(ep->d_name, l);
 
-                            debug_log("Adding candidate: %s/%s.", crash_dir, ep->d_name);
+                            DEBUGMSGTL(("rec:crash", "Adding candidate: %s/%s.", crash_dir, ep->d_name));
                         }
                     }
                 } else {
                     /* skip if there is no vmcore file */
-                    debug_log ("Skipping: %s (no vmcore found).", ep->d_name);
+                    DEBUGMSGTL(("rec:crash", "rec:crash", "Skipping: %s (no vmcore found).", ep->d_name));
                 }
             }
             closedir(dp);
@@ -792,7 +765,7 @@ static llist* find_crash_entries(int previous_flag) {
             perror ("Couldn't open crash directory.");
         }
     } else {
-        debug_log ("Crash directory does not exist (%s), skipping.", crash_dir);
+        DEBUGMSGTL(("rec:crash", "rec:crash", "Crash directory does not exist (%s), skipping.", crash_dir));
     }
     return l;
 }
@@ -837,7 +810,7 @@ static int update_previous_crash_entries(llist* cur, llist* prev) {
                 fprintf(fp, "This file notes that this crash has already been processed\n");
                 fprintf(fp, "and recorded by hpHelper.\n");
                 fclose(fp);
-                debug_log ("Created %s", filename);
+                DEBUGMSGTL(("rec:crash", "rec:crash", "Created %s", filename));
             }
         }
 
@@ -971,28 +944,28 @@ static int process_crash_item(char *entry) {
         memset(cmdbuf, 0, BUFSIZ);
         snprintf(cmdbuf, BUFSIZ-1, "%s --dump-dmesg %s/%s/vmcore %s",
                  dep.makedumpfile_bin, crash_dir, entry, dmesg_output);
-        debug_log("RUN: %s", cmdbuf);
+        DEBUGMSGTL(("rec:crash", "RUN: %s", cmdbuf));
         rc += run_proc_return_result(cmdbuf, NULL);
 
         if (is_gzip_binary_available()) {
             memset(cmdbuf, 0, BUFSIZ);
             snprintf(cmdbuf, BUFSIZ-1, "%s --best %s", dep.gzip_bin, dmesg_output);
-            debug_log("RUN: gzip --best %s", dmesg_output);
+            DEBUGMSGTL(("rec:crash", "RUN: gzip --best %s", dmesg_output));
             append_gz(dmesg_output, PATH_MAX);
             rc += run_proc_return_result(cmdbuf, NULL);
         }
 
-        debug_log("LOG: %s", dmesg_output);
+        DEBUGMSGTL(("rec:crash", "LOG: %s", dmesg_output));
         log_file_contents(dmesg_output);
 
         if (is_crash_binary_available()) {
-            debug_log("crash is available");
+            DEBUGMSGTL(("rec:crash", "crash is available"));
             if (is_debug_symbols_available()) {
                 char vmcore_loc[PATH_MAX];
                 char vmlinux_loc[PATH_MAX];
                 char release[PATH_MAX];
 
-                debug_log ("Debug symbols are available");
+                DEBUGMSGTL(("rec:crash", "Debug symbols are available"));
                 memset(vmcore_loc, 0, PATH_MAX);
                 memset(vmlinux_loc, 0, PATH_MAX);
                 memset(release, 0, PATH_MAX);
@@ -1010,38 +983,38 @@ static int process_crash_item(char *entry) {
                              dep.crash_bin, vmlinux_loc, vmcore_loc, 
                              crash_input_script, crash_output);
                     
-                    debug_log("RUN: %s", cmdbuf);
+                    DEBUGMSGTL(("rec:crash", "RUN: %s", cmdbuf));
                     rc += run_proc_return_result(cmdbuf, NULL);
 
                     if (is_gzip_binary_available()) {
                         memset(cmdbuf, 0, BUFSIZ);
                         snprintf(cmdbuf, BUFSIZ-1, "%s --best %s", dep.gzip_bin, crash_output);
-                        debug_log("RUN: gzip --best %s", crash_output);
+                        DEBUGMSGTL(("rec:crash", "RUN: gzip --best %s", crash_output));
                         append_gz(crash_output, PATH_MAX);
 
                         rc += run_proc_return_result(cmdbuf, NULL);
                     }
 
 
-                    debug_log ("LOG: %s", crash_output);
+                    DEBUGMSGTL(("rec:crash", "LOG: %s", crash_output));
                     rc = log_file_contents(crash_output);
                 } else {
-                    debug_log ("Failed to write crash input script.");
+                    DEBUGMSGTL(("rec:crash", "Failed to write crash input script."));
                     log_buffer_contents("Failed to write crash input script.");
                     rc = 1;
                 }
             } else {
-                debug_log ("Debug package matching the running kernel is not installed.");
+                DEBUGMSGTL(("rec:crash", "Debug package matching the running kernel is not installed."));
                 log_buffer_contents("Install the matching kernel debug package to enable the extraction of crash information.");
                 rc = 1;
             }
         } else {
-            debug_log ("The crash binary is not installed.");
+            DEBUGMSGTL(("rec:crash", "The crash binary is not installed."));
             log_buffer_contents("Install the crash package to enable the extraction of kdump information.");
             rc = 1;
         }
     } else {
-        debug_log ("The makedumpfile binary is not installed.");
+        DEBUGMSGTL(("rec:crash", "The makedumpfile binary is not installed."));
         log_buffer_contents("makedumpfile is not available on this system.  Cannot extract crash information.");
         rc = 1;
     }
@@ -1073,7 +1046,7 @@ static int process_new_crashes(llist* cur, llist* prev) {
             /* this is a new crash */
             rc += process_crash_item(tmp->entry);
         } else {
-            debug_log ("%s has already been processed before", tmp->entry);
+            DEBUGMSGTL(("rec:crash", "%s has already been processed before", tmp->entry));
         }
         tmp = tmp->next;
     }
@@ -1095,6 +1068,8 @@ void LOG_PROCESS_CRASHES(void) {
     int rc = 0;
     llist* current_crash_entries = NULL;
     llist* previous_crash_entries = NULL;
+    int fd;
+
     struct stat sb;
 
     /* enable debugging based on a trigger file, rather than command
@@ -1102,12 +1077,11 @@ void LOG_PROCESS_CRASHES(void) {
     memset(&sb, 0, sizeof(struct stat));
     if (does_exist(DEBUG_TRIGGER_FILE)) {
         debug = 1;
-        debug_log ("Found %s, enabling debugging", DEBUG_TRIGGER_FILE);
     }
-    
+
     memset(&dep, 0, sizeof(os_dep_config));
     get_os_deps(&dep);
-    debug_log ("Loaded OS dependencies.");
+    DEBUGMSGTL(("rec:crash", "Loaded OS dependencies."));
     
     /* List entries in crash directory */
     current_crash_entries = find_current_crash_entries();
@@ -1124,14 +1098,13 @@ void LOG_PROCESS_CRASHES(void) {
             /* update previous_crash_entries file */
             rc = update_previous_crash_entries(current_crash_entries, previous_crash_entries);
         } else {
-            debug_log("Failed to process one or more crashes.  Not updating crash list.");
+            DEBUGMSGTL(("rec:crash", "Failed to process one or more crashes.  Not updating crash list."));
         }
 
     } else {
-        debug_log("No crashes found.");
+        DEBUGMSGTL(("rec:crash", "No crashes found."));
     }
 
-    close_debug_log();
     free_list(current_crash_entries);
     free_list(previous_crash_entries);
     return;
